fix: staging scripts do not import from index (with side effects)
