+++ src/stagingScripts.ts
@@ -1,5 +1,5 @@
-import { processActions } from "./index.js";
 import { checkRequiredEnvVars } from "./requiredEnvVars.js";
+import { processActions } from "./processActions.js";
 
 console.log("S T A G I N G   S C R I P T S")
